backend/requirements.txt
Django>=4.2
djangorestframework>=3.15
djangorestframework-simplejwt>=2.9
django-cors-headers>=4.0
django-prometheus>=2.1
django-filter>=23.2
openpyxl>=3.1
psycopg2-binary>=2.9
